Extract graph dispatch into src/runner/graph.rs
handle_graph and write_graph_artefact moved into a focused submodule so
src/runner/mod.rs drops from 451 to 385 lines — back under the
AGENTS.md 400-line cap. load_manifest_with_stage_reporting is promoted
to pub(super) so the new sibling module can call it; behaviour is
otherwise unchanged.
